requirements.txt
pinecone-client>=6.0.0
python-dotenv>=1.0.0
